Required changes has been made
