--- src/context/UserAuthContext.jsx
@@ -1,68 +1,0 @@
-import { createContext, useContext, useEffect, useState } from "react";
-import {
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-  onAuthStateChanged,
-  signOut,
-  GoogleAuthProvider,
-  signInWithPopup,
-  RecaptchaVerifier,
-  getAuth,
-  signInWithPhoneNumber
-} from "firebase/auth";
-import { auth } from "../firebase";
-import { DeleteForeverSharp } from "@material-ui/icons";
-
-
-const userAuthContext = createContext();
-
-export function UserAuthContextProvider({ children }) {
-  const [user, setUser] = useState({});
-
-  function logIn(email, password) {
-    return signInWithEmailAndPassword(auth, email, password);
-  }
-  function signUp(email, password) {
-    return createUserWithEmailAndPassword(auth, email, password);
-  }
-  function logOut() {
-    return signOut(auth);
-  }
-  function googleSignIn() {
-    const googleAuthProvider = new GoogleAuthProvider();
-    return signInWithPopup(auth, googleAuthProvider);
-  }
-
-  function setUpRecaptcha(number){
-    const auth = getAuth();
-    const  reCaptchaVerifier = new RecaptchaVerifier("recaptcha-container",
-    {},
-    auth
-    );
-    reCaptchaVerifier.render();
-    return signInWithPhoneNumber(auth,number,reCaptchaVerifier); 
-  }
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (currentuser) => {
-      console.log("Auth", currentuser);
-      setUser(currentuser);
-    });
-
-    return () => {
-      unsubscribe();
-    };
-  }, []);
-
-  return (
-    <userAuthContext.Provider
-      value={{ user, logIn, signUp, logOut, googleSignIn,setUpRecaptcha}}
-    >
-      {children}
-    </userAuthContext.Provider>
-  );
-}
-export function useUserAuth() {
-  return useContext(userAuthContext);
-}
-export default useUserAuth;
